refactor(catalog): extrae validateSupplyAttributes compartido (create/edit Supply)
Elimina la duplicación del helper de validación de atributos entre CreateSupply
y EditSupply: una única función `validateSupplyAttributes(attributeRepo, ...)`
(fuente única del flujo resolver-esquema-efectivo → validar). Sin cambio de
comportamiento ni de constructores/DI; el repo se pasa como argumento.
